Added basic testing to JudgingInstances are created correctly
Judges should have at least the number of projects in the groups they
can judge or the minimum number of projects per judge, whichever is
lower.

Projects should have at least the number of judges that can judge the
group or the minimum number or judges per project, whichever is lower.