Simplified and corrected the method for removing excess instances
The function now iterates through instances for judges and projects
that exceed the specified bounds. A new query is done every time an
instance is deleted. This may end up being a performance concern.

The function now also takes responses into account. So an instance with
a response will not be deleted.

Still need to implement that type of check during the "remove non-
matching instances" phase.