Resolved issues with m2m models
After all my tests were passing, I discovered that incremental
assignment didn't work because m2m relations were saved after the main
model (the Judge). Because the m2m signals fire for each m2m model,
that was not a viable solution either.

Instead, the judge fires it's own signal on commit of the transaction
where it is being saved. This should ensure it is triggered after all
m2m models are saved.

If there is ever a context in which those relations are changed without
calling the judge's save method, that would be a case that isn't covered
here.

This change broke all of the tests because django's test case has its
own transaction. I had to convert some of the tests over to
Transaction Test Case, which took a while to sort out.